Configuração do Docker na API Management.Interest -aplicação do appsettings para ambiente de desenvolvimento -criação de propriedade para definir a moeda do valor de retorno -ajuste do Swagger para rodar em container Docker -integração da API com o Redis no docker-compose

+++ src/Management.Interest/Management.Interest.Infrastructure.Data.Query.Tests/Unit/Queries/v1/GetInterestRate/Mocks/Mappers/MapperMock.cs
@@ -12,9 +12,9 @@ namespace Management.Interest.Infrastructure.Data.Query.Tests.Unit.Queries.v1.Ge
         {
             IMapper mock = Substitute.For<IMapper>();
 
-            mock.Map<decimal?, GetInterestRateQueryResponse>(Arg.Is<decimal?>(interestRate => interestRate.HasValue)).Returns(GetInterestRateResponseSpecimen.For.ValidRequest);
+            mock.Map<double?, GetInterestRateQueryResponse>(Arg.Is<double?>(interestRate => interestRate.HasValue)).Returns(GetInterestRateResponseSpecimen.For.ValidRequest);
 
-            mock.When(x => x.Map<decimal?, GetInterestRateQueryResponse>(Arg.Is<decimal?>(interestRate => !interestRate.HasValue))).Do(response => { throw new NotFoundCustomException(); });
+            mock.When(x => x.Map<double?, GetInterestRateQueryResponse>(Arg.Is<double?>(interestRate => !interestRate.HasValue))).Do(response => { throw new NotFoundCustomException(); });
 
             return mock;
         }
